+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
-- 
+- Added spoilage season count config option
 ### Fixed:
-- Fixed farmers delight compat
+- 
 ### Changed:
 -